Turn error message container into a live regin
presently if someone were to cause a validation error with one of the date picker fields, error messages will appear at the bottom of the dialog.  However, as these messages are not presently injected into a live region, these error messages will not be exposed to screen reader users - failing wcag 4.1.3 Status Messages.

This PR adds an `aria-live=assertive` to the `tsi-errorMessageContainer` which will help ensure that these messages are announced when they dynamically appear.